# docs/openapi_v9.yaml
openapi: 3.0.3
info:
  title: Plan Charge v9 API
  version: 1.0.0
  description: |
    OpenAPI spec for Plan Charge v9 backend (FastAPI) covering identity, planning, integrations, and reporting.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    FromDate:
      name: from
      in: query
      schema: { type: string, format: date }
    ToDate:
      name: to
      in: query
      schema: { type: string, format: date }
  schemas:
...
    Organization:
      type: object
      required: [name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        timezone: { type: string, default: Europe/Paris }
        default_workweek:
          type: object
          additionalProperties: true
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
paths:
...
  /orgs:
    get:
      tags: [Orgs]
      summary: List organizations
      parameters: [ { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/PageSize' } ]
      responses:
        '200':
          description: List of orgs
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Organization' }
                  total: { type: integer }
    post:
      tags: [Orgs]
      summary: Create organization
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Organization' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Organization' }}}}
  /orgs/{id}:
    get:
      tags: [Orgs]
      summary: Get organization
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Organization' }}}}
    patch:
      tags: [Orgs]
      summary: Patch organization
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Organization' }
      responses:
        '200': { description: Updated }
    delete:
      tags: [Orgs]
      summary: Soft delete organization
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Deleted }
...
  /reports/utilization:
    get:
      tags: [Reports]
      summary: Utilization report
      parameters:
        - name: group_by
          in: query
          required: true
          schema: { type: string, enum: [team, person, project] }
        - { $ref: '#/components/parameters/FromDate' }
        - { $ref: '#/components/parameters/ToDate' }
      responses:
        '200': { description: OK }
  /reports/overbookings:
    get:
      tags: [Reports]
      summary: Overbooking detection
      parameters:
        - { $ref: '#/components/parameters/FromDate' }
        - { $ref: '#/components/parameters/ToDate' }
      responses:
        '200': { description: OK }
  /reports/capacity-vs-load:
    get:
      tags: [Reports]
      summary: Capacity vs load
      parameters:
        - name: team_id
          in: query
          schema: { type: string, format: uuid }
        - { $ref: '#/components/parameters/FromDate' }
        - { $ref: '#/components/parameters/ToDate' }
      responses:
        '200': { description: OK }
...